diera/pyproject.toml
[tool.poetry]
name = "diera"
version = "0.1.0"
description = "Website for the cultural center Diera do sveta"
authors = ["Marek Vician <marc.vician@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
easy-thumbnails = "^2.8.5"
html5lib = ">=1.0.1"
pillow = ">=3.0"
six = "^1.16.0"
pytz = "^2023.3"
python-dateutil = "^2.8.2"
whitenoise = "^6.4.0"
mysqlclient = "^2.1.1"
python-dotenv = "^1.0.0"
reportlab = "^3.6.12"
svglib = "^1.5.1"
humanreadable = "^0.2.0"
django = "3.2.3"
django-filer = ">=1.3"
django-treebeard = "4.4"
django-classy-tags = ">=2.0"
django-sekizai = ">=2.0"
django-mptt = ">0.9"
django-sortedm2m = "^3.1.1"
django-photologue = "^3.15.1"
django-environ = "^0.10.0"
django-cms = "3.9"
djangocms-admin-style = ">=2.0,<3.0"
djangocms-text-ckeditor = ">=4.0,<5.0"
djangocms-link = ">=3.0,<4.0"
djangocms-icon = ">=2.0,<3.0"
djangocms-style = ">=3.0,<4.0"
djangocms-googlemap = ">=2.0,<3.0"
djangocms-video = ">=3.0,<4.0"
djangocms-file = ">=3.0,<4.0"
djangocms-picture = ">=3.0,<4.0"
djangocms-bootstrap4 = ">=2.0,<3.0"
djangocms-snippet = "^3.1.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
bpython = "^0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
